Fix module system by taking out pprint functions.
Right now, pprint functions are defined using open variants on the constructors
used in both cast languages. This is problematic because this signature is hard
to match in the module type sig. For now, pprint is defined in the module Types
itself, and not inside each module encapsulating the POPL19 cast language and
the space-efficient SE cast language.